Freeze classes and modules more thoroughly (all attributes).
In order to make it fast enough, I had to memoize the freeze for
immutable objects, which meant I had to memoize the freeze for
immutable objects, which meant I had to know when objects are
immutable.

I also made the callgraph more precise when it uses only specific
attributes of a global (e.g. `module.function()`) or
parameter (e.g. `argument.method()`).